Remove nested class example in behavior context
The section on nested classes included a supposed bad example along with a good example of how to reflect nested classes in the behavior context. However, there was nothing syntactically wrong with the bad example. Using the supposed bad example in a test component proved this. The code compiled and worked fine, therefore the section has been removed.

Signed-off-by: William Hayward <wilhayw@amazon.com>