Update key-concepts.md
This link is broken. 
There is no replacement link, as you split Lua and Canvas scripting now, as opposed to how you did it at the time of this old link. 

Do you want to put a sentence in place of this one, that links to the Lua and the Canvas page?

I could do that for you, then. 

P.S: I strongly recommend introducing some CI check, that runs a (u)grep through the repo, to see if there are references left. 

Dead links don't need to be. 

Signed-off-by: ShalokShalom <shalokshalom@protonmail.ch>